tentative fix for discontinuity in undo history
when redoing the insertion of an element, a different QTreeWidgetItem is created. Subsequent redoings of elements or attributes below this one will fail because the original was remembered with creating the QUndoCommand. This fix does not solve this problem yet, but it doesn't do any damage either so let's keep it while we build the other actions